all interfaces moved to catalog named interfaces
