+++ src/components/navigation/navItems/Friends/SendFriendRequest.tsx
@@ -8,19 +8,8 @@ import {
   Grid,
   Typography,
 } from "@mui/material";
+import { SendFriendRequestProps } from "../../../../services/interfaces/restaurant";
 
-interface SendFriendRequestProps {
-  user: {
-    senderId: string;
-    senderName: string;
-  };
-  request: any;
-  isFriend: any;
-  isRequestReceived: any;
-  handleInvite: (userId: string) => void;
-  handleCancelInvite: (userId: string) => void;
-  handleRemoveFriend: (userId: string) => void;
-}
 
 const SendFriendRequest: React.FC<SendFriendRequestProps> = ({
   user,
