+++ src/components/restaurantManagement/restaurants/RestaurantData.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { RestaurantDataProps } from "../../../services/interfaces";
+import { RestaurantDataProps } from "../../../services/interfaces/restaurant";
 import EditSharpIcon from "@mui/icons-material/EditSharp";
 
 const RestaurantData: React.FC<RestaurantDataProps> = ({ restaurant }) => {
