+++ src/components/restaurantManagement/restaurants/restaurantsList/Group.tsx
@@ -5,7 +5,7 @@ import ListItemText from "@mui/material/ListItemText";
 import Collapse from "@mui/material/Collapse";
 import ExpandLess from "@mui/icons-material/ExpandLess";
 import ExpandMore from "@mui/icons-material/ExpandMore";
-import { GroupProps } from "../../../../services/interfaces";
+import { GroupProps } from "../../../../services/interfaces/restaurant";
 import { Button } from "@mui/material";
 
 const Group: React.FC<GroupProps> = ({
