+++ src/services/interfaces/navigation.tsx
@@ -1,0 +1,15 @@
+export interface NotificationListProps {
+    setHasNotifications: (hasNotifications: boolean) => void;
+  }
+  
+  export interface NotificationsButtonProps {
+    isDark: boolean;
+  }
+  
+  export interface ThemeButtonProps {
+    setIsDark: Function;
+  }
+  
+  export interface ToolsProps {
+    setIsDark: Function;
+  }
