+++ src/services/interfaces/register.tsx
@@ -1,0 +1,3 @@
+export interface RegisterEmpModalProps {
+    setIsModalOpen: Function;
+  }
